consumerNoSQL/requirements.txt
fdk
oci
kafka-python
requests